Create `EditableTextLine.cs`
